added card with desighn
