+++ website-react/src/index.css
@@ -5,6 +5,10 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  background-image: url("https://cdn-images-1.medium.com/max/800/0*ngXgBNNdx6iiWP8q.png");
+  background-repeat: no-repeat;
+  background-attachment: fixed;
+  background-size: cover;
 }
 
 code {
@@ -15,3 +19,11 @@ code {
   background-color: rgba(255, 165, 0, 0.67) !important;
 }
 /* '2' */
+
+.card-body {
+  box-shadow: 0px 0px 8px 3px #e2ae53;
+}
+
+.card-body i.fas.fa-sort-down {
+  float: right;
+}
